+++ src/platform/encounters/index.ts
@@ -1,3 +1,4 @@
 export { loadEncounterCatalogue } from './discovery'
 export { validateEncounterPackage } from './validate'
+export * from './actions'
 export type * from './types'